% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageBrowser.R
\name{imageBrowser}
\alias{imageBrowser}
\title{Shiny application to visualise multi-channel images}
\usage{
imageBrowser(
  image,
  mask = NULL,
  object = NULL,
  cell_id = NULL,
  img_id = NULL,
  ...
)
}
\arguments{
\item{image}{(optional) a \code{\linkS4class{CytoImageList}} object containing
single or multi-channel \code{\linkS4class{Image}} objects.}

\item{mask}{(optional) a \code{\linkS4class{CytoImageList}} containing
single-channel \code{\linkS4class{Image}} objects.}

\item{object}{a \code{\linkS4class{SingleCellExperiment}} object.}

\item{cell_id}{character specifying the \code{colData(object)} entry, in which
the integer cell IDs are stored. These IDs should match the integer pixel
values in the segmentation mask object (\code{mask}).}

\item{img_id}{character specifying the \code{colData(object)} and
\code{mcols(mask)} and/or \code{mcols(image)} entry, 
in which the image IDs are stored.}

\item{...}{parameters passed to the \code{\link{plotCells}} or
\code{\link{plotPixels}} function.}
}
\value{
A Shiny app object for multi-channel image visualization
}
\description{
TODO
}
\examples{
TODO


}
